+++ llvm-hs-pure/src/LLVM/AST/Typed.hs
@@ -33,6 +33,7 @@ instance Typed C.Constant where
   typeOf (C.Int bits _)  = IntegerType bits
   typeOf (C.Float t) = typeOf t
   typeOf (C.Null t)      = t
+  typeOf (C.AggregateZero t) = t
   typeOf (C.Struct {..}) = StructureType isPacked (map typeOf memberValues)
   typeOf (C.Array {..})  = ArrayType (fromIntegral $ length memberValues) memberType
   typeOf (C.Vector {..}) = VectorType (fromIntegral $ length memberValues) $